requirements.txt
numpy
pandas
matplotlib
absl-py
seaborn
scikit-learn
keras
tensorflow
tabulate
